+++ CHANGELOG.rst
@@ -15,6 +15,15 @@ Unreleased
 ----------
 * Configuration for automatic filters docs generation.
 
+[1.11.0] - 2024-09-30
+--------------------
+
+Added
+~~~~~
+
+* CourseAboutPageURLRequested filter added which can be used to modify the URL for the course about organization.
+* LMSPageURLRequested filter added which can be used to modify the URL for the LMS page organization.
+
 [1.10.0] - 2024-09-20
 --------------------